Move get_chords() method to Voicing
